Splitted ExchangeRate::BitcoinAdapter and ExchangeRate::Adapter
