+++ lib/straight.rb
@@ -30,14 +30,15 @@ require_relative 'straight/blockchain_adapters/insight_adapter'
 require_relative 'straight/blockchain_adapters/chain_com_adapter'
 
 require_relative 'straight/exchange_rate_adapter'
-require_relative 'straight/exchange_rate_adapters/bitpay_adapter'
-require_relative 'straight/exchange_rate_adapters/coinbase_adapter'
-require_relative 'straight/exchange_rate_adapters/bitstamp_adapter'
-require_relative 'straight/exchange_rate_adapters/localbitcoins_adapter'
-require_relative 'straight/exchange_rate_adapters/okcoin_adapter'
-require_relative 'straight/exchange_rate_adapters/btce_adapter'
-require_relative 'straight/exchange_rate_adapters/kraken_adapter'
-require_relative 'straight/exchange_rate_adapters/average_rate_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/bitpay_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/coinbase_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/bitstamp_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/localbitcoins_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/okcoin_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/btce_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/kraken_adapter'
+require_relative 'straight/exchange_rate_bitcoin_adapters/average_rate_adapter'
 
 require_relative 'straight/address_providers/bip32'
