+++ lib/straight/exchange_rate_adapter.rb
@@ -1,8 +1,6 @@
 module Straight
   module ExchangeRate
-
     class Adapter
-
       include Singleton
 
       class FetchingFailed       < StraightError; end
@@ -10,16 +8,6 @@ module Straight
 
       def initialize(rates_expire_in: 1800)
         @rates_expire_in = rates_expire_in # in seconds
-      end
-
-      def convert_from_currency(amount_in_currency, btc_denomination: :satoshi, currency: 'USD')
-        btc_amount = amount_in_currency.to_f/rate_for(currency)
-        Satoshi.new(btc_amount, from_unit: :btc, to_unit: btc_denomination).to_unit
-      end
-
-      def convert_to_currency(amount, btc_denomination: :satoshi, currency: 'USD')
-        amount_in_btc = Satoshi.new(amount.to_f, from_unit: btc_denomination).to_btc
-        amount_in_btc*rate_for(currency)
       end
 
       def fetch_rates!
@@ -47,7 +35,7 @@ module Straight
           if intermediate.respond_to?(:[])
             intermediate[key]
           else
-            raise CurrencyNotSupported 
+            raise CurrencyNotSupported
           end
         end
       end
@@ -57,8 +45,6 @@ module Straight
       def rate_to_f(rate)
         rate ? rate.to_f : raise(CurrencyNotSupported)
       end
-
     end
-
   end
 end
