+++ lib/straight/exchange_rate_bitcoin_adapters/bitstamp_adapter.rb
@@ -1,8 +1,8 @@
 module Straight
   module ExchangeRate
 
-    class BitstampAdapter < Adapter
-      
+    class BitstampAdapter < BitcoinAdapter
+
       FETCH_URL = 'https://www.bitstamp.net/api/ticker/'
 
       def rate_for(currency_code)
