+++ lib/straight/exchange_rate_bitcoin_adapters/btce_adapter.rb
@@ -1,13 +1,13 @@
 module Straight
   module ExchangeRate
 
-    class OkcoinAdapter < Adapter
-      
-      FETCH_URL = 'https://www.okcoin.com/api/ticker.do?ok=1'
+    class BtceAdapter < BitcoinAdapter
+
+      FETCH_URL = 'https://btc-e.com/api/2/btc_usd/ticker'
 
       def rate_for(currency_code)
         super
-        raise CurrencyNotSupported if currency_code != 'USD'
+        raise CurrencyNotSupported if !FETCH_URL.include?("btc_#{currency_code.downcase}")
         rate = get_rate_value_from_hash(@rates, 'ticker', 'last')
         rate_to_f(rate)
       end
