+++ lib/straight/exchange_rate_bitcoin_adapters/kraken_adapter.rb
@@ -1,8 +1,8 @@
 module Straight
   module ExchangeRate
 
-    class KrakenAdapter < Adapter
-      
+    class KrakenAdapter < BitcoinAdapter
+
       FETCH_URL = 'https://api.kraken.com/0/public/Ticker?pair=xbtusd'
 
       def rate_for(currency_code)
