+++ lib/straight/exchange_rate_bitcoin_adapters/okcoin_adapter.rb
@@ -1,13 +1,13 @@
 module Straight
   module ExchangeRate
 
-    class BtceAdapter < Adapter
-      
-      FETCH_URL = 'https://btc-e.com/api/2/btc_usd/ticker'
+    class OkcoinAdapter < BitcoinAdapter
+
+      FETCH_URL = 'https://www.okcoin.com/api/ticker.do?ok=1'
 
       def rate_for(currency_code)
         super
-        raise CurrencyNotSupported if !FETCH_URL.include?("btc_#{currency_code.downcase}")
+        raise CurrencyNotSupported if currency_code != 'USD'
         rate = get_rate_value_from_hash(@rates, 'ticker', 'last')
         rate_to_f(rate)
       end
