+++ spec/lib/exchange_rate_bitcoin_adapter_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-RSpec.describe Straight::ExchangeRate::Adapter do
+RSpec.describe Straight::ExchangeRate::BitcoinAdapter do
 
   class Straight::ExchangeRate::Adapter
     FETCH_URL = ''
   end
 
   before(:each) do
-    @exchange_adapter = Straight::ExchangeRate::Adapter.instance
+    @exchange_adapter = Straight::ExchangeRate::BitcoinAdapter.instance
   end
 
   describe "converting currencies" do
