+++ spec/lib/exchange_rate_bitcoin_adapters/average_rate_adapter_spec.rb
