+++ spec/lib/exchange_rate_bitcoin_adapters/bitpay_adapter_spec.rb
