+++ spec/lib/exchange_rate_bitcoin_adapters/bitstamp_adapter_spec.rb
