+++ spec/lib/exchange_rate_bitcoin_adapters/btce_adapter_spec.rb
