+++ spec/lib/exchange_rate_bitcoin_adapters/coinbase_adapter_spec.rb
