+++ spec/lib/exchange_rate_bitcoin_adapters/kraken_adapter_spec.rb
