+++ spec/lib/exchange_rate_bitcoin_adapters/localbitcoins_adapter_spec.rb
