+++ spec/lib/exchange_rate_bitcoin_adapters/okcoin_adapter_spec.rb
